Fixed routing and moved dashboard to new folder
