+++ frontend/src/App.js
@@ -1,48 +1,33 @@
-import React, { Component } from 'react';
-import { Switch, Route } from 'react-router-dom';
+import './App.css';
+
+import React from 'react';
+import {Switch, Redirect, Route} from 'react-router-dom';
+
 import Login from './Login';
 import Registration from './Registration';
-import Dashboard from './teacher/classes.js'
-import TeacherCalendar from './teacher/calendar.js'
-import logo from './logo.svg';
-import './App.css';
+import Dashboard from './Dashboard';
+import TeacherCalendar from './teacher/calendar.js';
 import StudentClass from './student/classes.js';
 import TeacherFeedback from './teacher/feedback.js';
 
-class App extends Component {
-  previousLocation = this.props.location;
-
-  componentWillUpdate(nextProps) {
-    const { location } = this.props;
-    // set previousLocation if props.location is not modal
-    if (
-      nextProps.history.action !== "POP" &&
-      (!location.state || !location.state.modal)
-    ) {
-      this.previousLocation = this.props.location;
-    }
-  }
-
-  render() {
-    const { location } = this.props;
-    const isModal = !!(
-      location.state &&
-      location.state.modal &&
-      this.previousLocation !== location
-    ); // not initial render
-    return (
-      <div>
-          <Switch location={isModal ? this.previousLocation : location}>
-            <Route exact path="/" component={Login} />
-            <Route path="/registration" component={Registration} />
-            <Route path="/dashboard" component={Dashboard} />
-            <Route path="/teacher/calendar" component={TeacherCalendar} />
-            <Route path="/student/classes" component={StudentClass} />
-            <Route path="/teacher/feedback" component={TeacherFeedback} />
-          </Switch>
-      </div>
+const App = props => {
+  const handleInvalidURL = () =>
+    localStorage.length === 0 ? (
+      <Redirect to="/login" />
+    ) : (
+      <Redirect to="/dashboard" />
     );
-  }
-}
+  return (
+    <Switch>
+      <Route path="/login" component={Login} />
+      <Route path="/registration" component={Registration} />
+      <Route path="/dashboard" component={Dashboard} />
+      <Route path="/teacher/calendar" component={TeacherCalendar} />
+      <Route path="/student/classes" component={StudentClass} />
+      <Route path="/teacher/feedback" component={TeacherFeedback} />
+      <Route path="*" render={handleInvalidURL} />
+    </Switch>
+  );
+};
 
 export default App;
