+++ frontend/src/index.js
@@ -2,17 +2,18 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
-import { BrowserRouter as Router, Route} from 'react-router-dom';
+import {BrowserRouter as Router, Route} from 'react-router-dom';
 import * as serviceWorker from './serviceWorker';
-import { CookiesProvider } from 'react-cookie';
+import {CookiesProvider} from 'react-cookie';
 
 ReactDOM.render(
-      <CookiesProvider>
-        <Router>
-          <Route component={App} />
-        </Router>
-      </CookiesProvider>
-    , document.getElementById('root'));
+  <CookiesProvider>
+    <Router>
+      <Route component={App} />
+    </Router>
+  </CookiesProvider>,
+  document.getElementById('root'),
+);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
